Part 2 but it's too slow

+++ src/day13/ILPSolver.vb
@@ -1,16 +1,16 @@
 Namespace AdventOfCode
 Public Class ILPSolver
-    Public Shared Function Solve(objectiveFunction As Func(Of Integer, Integer, Double),
-                                  constraintFunction As Func(Of Integer, Integer, Boolean),
-                                  maxNA As Integer,
-                                  maxNB As Integer) As (Integer, Integer)
+    Public Shared Function Solve(objectiveFunction As Func(Of Double, Double, Double),
+                                  constraintFunction As Func(Of Double, Double, Boolean),
+                                  maxNA As Double,
+                                  maxNB As Double) As (Double, Double)
         Dim minCost As Double = Double.MaxValue
-        Dim bestNA As Integer = 0
-        Dim bestNB As Integer = 0
+        Dim bestNA As Double = 0
+        Dim bestNB As Double = 0
 
         ' Exhaustive search for nA and nB
-        For nA As Integer = 0 To maxNA
-            For nB As Integer = 0 To maxNB
+        For nA As Double = 0 To maxNA
+            For nB As Double = 0 To maxNB
                 ' Check feasibility
                 If constraintFunction(nA, nB) Then
                     ' Calculate cost
